AlexeiSorokin/django-jobs-base-master/requirements.txt
alembic==0.9.2
amqp==2.1.4
appdirs==1.4.3
asn1crypto==0.22.0
attrs==16.3.0
Automat==0.5.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.3.2
billiard==3.5.0.2
bleach==2.0.0
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
configparser==3.5.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==1.8.1
cssselect==0.9.1
decorator==4.0.11
django==1.11.1
django-extensions==1.7.9
django-class-based-auth-views==0.4
entrypoints==0.2.2
enum34==1.1.6
feedfinder2==0.0.1
feedparser==5.1.3
functools32==3.2.3-2
geograpy==0.3.7
html5lib==1.0b8
idna==2.5
incremental==16.10.1
ipaddress==1.0.18
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipython-sql==0.3.8
ipywidgets==6.0.0
jellyfish==0.5.1
jieba==0.35
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
kombu==4.0.2
lxml==3.3.5
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
newspaper==0.0.9.8
nltk==3.2.2
notebook==5.0.0
numpy==1.12.1
packaging==16.8
pandocfilters==1.4.1
parsel==1.1.0
pathlib2==2.2.1
pbr==3.0.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.5.1
prettytable==0.7.2
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycountry==17.1.8
pycparser==2.17
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.4.0
python-telegram-bot==5.3.1
pytz==2017.2
PyYAML==3.11
pyzmq==16.0.2
qtconsole==4.3.0
queuelib==1.4.2
raven==6.0.0
requests==2.3.0
scandir==1.5
Scrapy==1.3.3
scrapyd==1.2.0
scrapyd-client==1.1.0
scrapy-sentry==0.7.0
scrapyscript==0.0.6
selenium==3.4.2
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.9
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
terminado==0.6
testpath==0.3
tldextract==1.5.1
tornado==4.5.1
traitlets==4.3.2
Twisted==17.1.0
vine==1.1.3
w3lib==1.17.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
zope.interface==4.3.3
